Create Code.ino
Two objects created, one Keypad and an LCD. We keep on getting the input from the Keypad and as long as we don't get an operator pressed then we ask for the second number by calling the method defined at the bottom. that function only returns the second number when the equals button is pressed. After that, we calculate according to the required operator and print on the second line.